+++ v2/src/core/config/AppConfig.js
@@ -7,7 +7,7 @@
   global.AppConfig = Object.freeze({
     app: Object.freeze({
       name: "My Produce Assistant",
-      version: "51.1"
+      version: "51.2"
     }),
 
     urls: Object.freeze({